Implement "median of medians" algorithm
Use the "median of median" algorithm in
median_quickselect.c to guarantee a linear worst case
performance (compared to the previous N^2 worst case).

I also, in comments, added an implementation for
random pivot selection - I have not done proper
profiling yet, but in my un-scientific testing on my
personal machine, its performance was slightly better
on average than the "median of medians" implementation,
with about 1/10 of the code.